break NetworkCLIENT* methods out into separate file
